revert inline
